refactor: rename to Auth
