+++ frontend/src/components/Auth/auth0-provider-with-navigate.tsx
