+++ frontend/src/main.tsx
@@ -7,7 +7,7 @@ import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
 
 import App from './App.tsx';
-import { Auth0ProviderWithNavigate } from './components/Auth0/auth0-provider-with-navigate.tsx';
+import { Auth0ProviderWithNavigate } from './components/Auth/auth0-provider-with-navigate.tsx';
 import { client } from './graphql/client.ts';
 import { store } from './store.ts';
